13:27 jaime

+++ classes/Pilot.java
@@ -2,7 +2,7 @@ package proyecto.classes;
 
 public class Pilot extends Employee{
   public int salary=100000;
-  public int hoursWorked;
+  public float hoursWorked;
   public String work="pilot";
 
   public Pilot(){
@@ -12,7 +12,7 @@ public class Pilot extends Employee{
   String date, String nation){
     super(dni, name, apellido, date, nation);
   }
-  public void addHours(int h){
+  public void addHours(float h){
     hoursWorked+=h;
     salary+=5000;
   }
