+++ classes/Tripulation.java
@@ -2,7 +2,7 @@ package proyecto.classes;
 
 public class Tripulation extends Employee{
   public int salary=40000;
-  public int hoursWorked;
+  public float hoursWorked;
   public String work="steward";
 
   public Tripulation(){
@@ -12,7 +12,7 @@ public class Tripulation extends Employee{
   String date, String nation, int sala){
     super(dni, name, apellido, date, nation);
   }
-  public void addHours(int h){
+  public void addHours(float h){
     hoursWorked+=h;
     salary+=1000;
   }
